# Optimize GCP provisioning speed: bundle all runtime file_mounts.