File mounting our runtime files hurts provisioning speed. This PR bundles them into 1 file mount. Copying them locally takes 10-20ms.

Currently for GCP only. Will be ported to other clouds later.


Before (taken from #1092) - median 2:40
```
sky launch -c opt --cloud gcp -y ''  13.89s user 3.17s system 10% cpu 2:43.79 total
sky launch -c opt2 --cloud gcp -y ''  11.78s user 2.37s system 8% cpu 2:40.32 total
sky launch -c opt9 --cloud gcp -y ''  11.49s user 2.37s system 9% cpu 2:31.00 total
```

After - median 2:05 (22% speedup)
```
sky launch -c opt6 --cloud gcp -y ''  12.54s user 2.62s system 12% cpu 2:05.92 total
sky launch -c opt7 --cloud gcp -y ''  11.73s user 2.29s system 11% cpu 2:02.68 total
sky launch -c opt8 --cloud gcp -y ''  10.92s user 2.06s system 10% cpu 2:05.78 total
```

Breakdown of the current overheads, updating the one recorded in #1092 (main change is the file mount part):
Total: ~2min 5s
- `_add_auth_to_cluster_config()`: ~1.5-2s
- **Launch VM till SSH ready: ~55s-1min**
  - Includes: check head node first + provision & wait for ssh + update status
  - [ ] Note that ray autoscale's "update status" using tags is expensive, 4-7s
- **Preparing SkyPilot runtime: ~1min 5s**
  - **File mount our runtime files bundle: ~5s**
    - see the timestamp between `[2/7] Processing file mounts` to `[3/7] No worker file mounts to sync`
  - **Setup_commands: ~31s**
    - between `[6/7] Running setup commands` and `[7/7] Starting the Ray runtime`
   - **Starting Ray: ~16s**
     - between `[7/7] Starting the Ray runtime` and last line of provision.log
 - some other overheads (e.g., update-status 6 times!)
```
» grep '\/7\]'  /Users/zongheng/sky_logs/sky-2022-08-19-08-36-20-887250/provision.log    2 ↵
2022-08-19 08:36:47,959 INFO updater.py:241 -- [1/7] Waiting for SSH to become available
2022-08-19 08:37:24,795 INFO updater.py:219 -- [2/7] Processing file mounts
2022-08-19 08:37:29,605 INFO updater.py:236 -- [3/7] No worker file mounts to sync
2022-08-19 08:37:36,038 INFO updater.py:397 -- [4/7] No initialization commands to run.
2022-08-19 08:37:36,038 INFO updater.py:401 -- [5/7] Initalizing command runner
2022-08-19 08:37:36,038 INFO updater.py:410 -- [6/7] Running setup commands
2022-08-19 08:38:07,074 INFO updater.py:450 -- [7/7] Starting the Ray runtime
```

Tested
- [x] smoke tests except the 4 known failures documented in #1092